home: Use weightVariableTextStyle instead of TextStyle.fontWeight
Now, this text is styled with a `wght` value of 700, instead of 400
from the ambient style. The FontWeight.bold seems to have already
been causing the spans to render more boldly than surrounding text,
but in a way that doesn't match what we get for Source Code Pro with
`wght` 700. I think this is because of
  https://github.com/flutter/flutter/issues/136779
.

We could probably have left out the `const TextStyle().merge()`; I
think the only difference it causes is in the resulting style's
debugLabel. But I figured we might as well be consistent in how we
use weightVariableTextStyle, and in any case it should disappear
with
  https://github.com/flutter/flutter/issues/148026 .