content: Use weightVariableTextStyle in ContentTheme.textStyleError
This converts the last remaining place where we were using
TextStyle.fontWeight directly, when we should have been using
weightVariableTextStyle.

(In a future with
  https://github.com/flutter/flutter/issues/148026
, though, we'll want to change all uses of weightVariableTextStyle
back to plain TextStyle.fontWeight.)

Now, this text is styled with a `wght` value of 700, instead of 400
from the ambient style. The FontWeight.bold seems to have already
been causing the spans to render more boldly than surrounding text,
but in a way that doesn't match what we get for Source Code Pro with
`wght` 700. I think this is because of
  https://github.com/flutter/flutter/issues/136779
.