lightbox: Apply correct fix for slider flickering
This changes makes it so that the async function `seekTo` is awaited
before the `setState` in `Slider.onChangeEnd`, thus first setting the
controller's value which as a side-effect calls the `setState` in
`_handleVideoControllerUpdate`, and then the `setState` in
`Slider.onChangeEnd`.

Resulting in both setState calls to be coalesced in a single redraw.
Whereas previously the unawaited `seekTo` call would delay setting
the controller's value in a microtask.