deps: Update pigeon to latest
Changelog:
  https://pub.dev/packages/pigeon/changelog

The only breaking change is:

  Limits the number of total custom types to 126.

  If more than 126 custom types are needed, consider breaking up
  your definition files.

Which definitely isn't a problem for us!

The changes all look pretty innocuous; but just to confirm, I fired
up the app (as a release build) and tested end-to-end that
notifications still appear.  As usual I tested it while the app was
in the foreground, in the background, and not running.