text: Add and use localizedTextBaseline
This seems like a handy function to have around; there were a
handful of places that were wanting something like this.

In b75908bef, we aimed to make ContentTheme.textStylePlainParagraph
a very direct, NFC translation of some Typography defaults, but we
accidentally fell short a little bit: I had forgotten that we
support Japanese, and that commit just hardcoded
TextBaseline.alphabetic, which was a regression. Since the "geometry
TextThemes" for M3 only differ in `textBaseline` [1], I believe that
commit would have been a pure translation if we had used this new
function. (We did follow up with 6284ed552, removing the
textBaseline field altogether. But that put the style at odds with
text basically everywhere else, which inherits a textBaseline from
Typography.)

I was hoping to see an improvement in how Japanese text is rendered
after this change, but I wasn't able to; I didn't notice any change
in appearance between TextBaseline.alphabetic and
TextBaseline.ideographic (or a TextStyle that omits the field, in
the case of ContentTheme.textStylePlainParagraph.) Anyway, if this
is because of incomplete support for `ideographic`, at least we'll
be well-prepared to take improvements as they arrive.

Someone did post some screenshots with both values rendering
identically:
  https://github.com/flutter/flutter/issues/94680

[1] The Material 2 spec apparently called for more differences that
    were removed in Material 3; see the docs for
    Typography.dense2021 and tall2021.