msglist: Tweak app-bar title to look better after recent text-size change
In a recent commit in this series, we decreased the app-bar title
text size, to align with the new Figma.

The spacing and alignment in MessageListAppBarTitle for stream and
topic narrows looked a little off after doing that. This should make
it look better. I'm leaving the comment saying it's "a bit ad hoc",
though; this will remain true until we have a Figma design to
follow.