subscription_list test: Set up whole ZulipApp, not just MaterialApp
Just like we did for the InboxPage tests, in the previous commit.

Here, there's one test that failed on the assumption that a
particular icon was the only one in view; it apparently isn't
anymore (probably because of the page's back button), and that
assumption isn't important to the goal of the test. So, we adapt by
removing that assumption.

Related: #393