api [nfc]: Rename Message.topic, from "subject"
The name "subject" is long obsolete for Zulip topics.  It persists
in a few places in the API, but we should keep it out of our code.
We had it here on the Message type really just because this is a
part of the code that I wrote in the very first hours of the
prototype of this app, before I'd read up enough on the
json_serialization package to find `JsonKey.name`.

There's a TODO referring to a few other similar names in the API;
we'll handle those shortly.