message: Update editState on events.
When a channel (a.k.a stream) or a topic gets updated on a message, we
get a UpdateMessageEvent from the server. The same happens when the
message's content is updated.

This updates the edit state to show the "edited" or "moved" marker
properly on the messages affected, and leave other move related states
(like streamId) untouched until we get to #150.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>