star: Align the icon with message content.
By wrapping the icon with a dummy Text widget in a Row, the Text helps
us stretch the row to the same height as a line of a message content
would have. This allows us to position the icon with respect to text,
even if a scaling factor has been applied.

Fixes #616.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>