message: Handle message deletion event.
When handling a message deletion event, we just simply remove the
messages from the per account store and wipe them from the message list
view if they are present.

Because message deletion can affect what items are to be seen (date
separator, recipient header, etc.), if the view has any message deleted, a
`_reprocessAll` call is triggered. It would be ideal to just removed the
affected items for better performance.

Fixes #120.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>