fake_api [nfc]: Add `delay` option to connection.prepare
This allows simulating a request that takes some time to return
and consequently races with something else.  For example here:
  https://github.com/zulip/zulip-flutter/pull/713#discussion_r1657899344

This is NFC because `Future.delayed(Duration.zero, f)` is exactly
equivalent to `Future(f)`.  (The docs aren't real clear on this;
but reading the implementations confirms they boil down to the
very same `Timer` constructor call, and then the docs do seem to be
trying to say that when read in light of that information.)