test: Generate random increasing streamId for example data by default
This allows `eg.stream` to be used in contexts where it's important
that the stream IDs don't collide or even that they're increasing,
without the test having to explicitly give specific stream IDs when
no other fact about them is relevant.

Some tests are dependent on the stream IDs being static. Those are
updated to use `123` as the value of `eg.stream` explicitly.