autocomplete [nfc]: Factor out AutocompleteDataProvider
`AutocompleteDataProvider` provides the full list of possible items
and filter predicate for `AutocompleteView` without having to deal
with state management or result computation process.

With this factored out of `AutocompleteView` it's responsibility is
limited to only managing the autocomplete state.

While fetching data and setting the logic on passing items to query
filter is the responsibility of `AutocompleteDataProvider`.