message [nfc]: Pull out orig/new values from move as local variables
This lets the Dart compiler see that these are definitely constant
through the function (which it can't assume when they're getters on
an object).  That in turn lets it narrow the types based on our
null checks, so that we can leave out `!` operators later.

That's small now, but will be more useful as we add more logic to
this function -- if we had to scatter `!` all around that logic,
we'd have to wonder whether some of them were unjustified assumptions
and might throw at runtime.

Also rename one local to "newStreamId", matching the property it
mirrors.  We'll rename both the variable and the property together,
as part of a sweep coming up soon (#631).