android: Cut unused network-security-config XML
This was introduced in cffb112fd / #474 because we thought it
would be a fix for #461.  As discussed on that issue thread,
though, it turns out this file doesn't have an effect on the
Dart HTTP implementation, which we use.

Since it looks like it's doing something but in fact it isn't,
cut it out to avoid confusion.  It'll still be there in the
Git history if we later want to use it as part of a future
effort to fix #461.