test: Add TestZulipApp; replace uses of ZulipApp
Greg noticed an instance where using ZulipApp directly was causing
tests to slow down by some 75%:
  https://github.com/zulip/zulip-flutter/pull/805#discussion_r1675335632
Specifically, that was when we converted the emoji-reaction widget
tests over to use ZulipApp instead of mimicking it with a
MaterialApp, etc.

It's helpful to use something approaching the real ZulipApp. But
we'd like to avoid that slowdown, especially if it's happening in
code that most tests don't need to exercise. So, here's a
stripped-down version that does that. It takes roughly the same time
as the MaterialApp approach in those emoji-reaction tests. (11s on
my machine, vs. 16s with ZulipApp.)

This commit just converts the tests that were using ZulipApp. Next,
we'll sweep through and treat the ones using the ad hoc MaterialApp
approach.