channel: Split up asserts in SubscriptionAddEvent handler
I just hit one of these in a new test I was writing, and wanted
more information to debug: which of the two conditions failed?

I seem to recall hearing a report of this in a non-test context
recently, too, but I'm not finding it on a quick search of the
tracker or of Zulip.