actions [nfc]: Move markNarrowAsRead to new file
This function isn't quite part of the UI -- it's mostly a driver
around some API requests.  But it doesn't belong in the API code
either, because it interacts with UI widgets in order to give
progress and failure feedback to the user.

Because there's nontrivial logic in here and it isn't really specific
to this particular piece of UI, we'd like to reuse much of this code,
in particular for #131 marking as unread (as discussed at #779).
This new file provides an appropriate home for that reusable code.

Before we can actually reuse this for marking as unread, it'll need
some refactoring.  Before we undertake refactoring it, it should have
tests that operate at the layer of this function itself, so that they
can easily be adapted to cover the mark-unread case too.  There are
existing tests that cover this pretty thoroughly, but they act on the
UI in the message list that drives it.  So we'll port those tests over
in the next couple of commits.