actions test: Port most MarkAsReadWidget tests to markNarrowAsRead
This sets us up to be able to refactor this function for other
use cases that aren't exercised by MarkAsReadWidget, and have good
tests for those cases too.

There are a couple of pieces of functionality of the widget's
_handlePress method that probably should be within the function
markNarrowAsRead instead, so that they can be reused together.
For the tests for those, we port them too but mark them as
skipped, with TODO comments.  After refactoring to move that
functionality inside, we can enable the tests.

The new markNarrowAsRead unit tests do most of the work of tests
for MarkAsReadWidget.  We leave behind a couple of smoke tests
that mostly just show it is indeed using markNarrowAsRead, plus
the tests whose ported versions are skipped.