submessage: Add SubmessageData classes for polls.
The sealed class `SubmessageData` is not actually in use, we could
potentially implement a discriminator utilizing the sealed class to
deserialize individual submessage content, but it is far easier to do so
when we have access to the full list of submessages. `SubmessageData` is
there for self-documentation purposes.

It is also worth noting that much of these class definitions are based
on previous reverse engineering effort and the web implementation. See:

- https://github.com/zulip/zulip/blob/304d948416465c1a085122af5d752f03d6797003/web/shared/src/poll_data.ts
- https://github.com/zulip/zulip-mobile/blob/2217c858e207f9f092651dd853051843c3f04422/src/api/modelTypes.js#L800-L861

Due to the flexibility of the submessage API, these classes tend to be
intentionally defensive against unknown or invalid values.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>