recent_senders test [nfc]: Cut unneeded IDs; make names look more different
Most of these tests don't use the specific stream IDs, user IDs,
or message IDs, so cut them out to reduce the number of details
for the reader to read.

Then also rename the streams, topics, and users so that they're
a bit more visually distinct from each other.  That helps in seeing
what's going on in the tests that are all about which bits of data
are the same as, or different from, which other ones.