api: Construct polls data store from submessages.
For now, we only consider the case when the submessages describe a
poll, and disgard them otherwise. It will be a simple refactor later to
support other Zulip widgets like todo lists, by extracting a common
ancestors for such widget data structures.

The `Poll` data structure will become useful when we support submessage
events, where updates to a poll can happen.

For testing, [Poll.submessages] is used for implementing the custom
toJson method. Because most of the messages do not have a poll, storing
a copy of submessages shouldn't be costy.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>