api: Construct polls data store from submessages.
For now, we only consider the case when the submessages describe a
poll, and disgard them otherwise. It will be a simple refactor later to
support other Zulip widgets like todo lists, by extracting a common
ancestors for such widget data structures.

The `Poll` data structure will become useful when we support submessage
events, where updates to a poll can happen.

Some more comments are added here instead of earlier because of their
references to `SubmessageData`.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>