recent_senders [nfc]: Rename some foosByBar that mean foosInBar
A map of "foos by bar" is a map where each key is a "bar", and
each value is a "foo" -- the phrase can be read as an abbreviation
of "foos indexed by bar" or "foos, where each foo is indexed by
its respective bar".

So in this function `messagesByUser` is accurately named because
a key represents a user and a value represents some messages.

But these other variables should be e.g. `topicsInStream`, not
`topicsByStream`.  That one isn't a map where the key is a stream
and the value is a topic or topics; rather it's a collection of
topics (and more data about them) which is the collection
specifically of topics in the stream we're acting on.