autocomplete test: Cut tests that could flake due to indeterminate sort
The [List.sort] method is documented as being not necessarily stable,
meaning that elements that compare equal could end up in an arbitrary
order in the result.  Here, that means that users which aren't
distinguished by recency in the topic or stream or in DM conversations
can appear in any order in the autocomplete results.

Several of these test cases were therefore vulnerable to flaking
because users 2 and 4 have no DMs and (in these tests) no messages
in the stream or topic.  I think what they cover is now also covered
by the "ranking across signals" tests above, together with the tests
for RecentSenders and RecentDmConversations that check those data
structures' handling of events.  So just cut the tests.