autocomplete test: Fix test that almost can't fail, of CombinedFeedNarrow
If for example one substitutes some other narrow like `topicNarrow` in
this `checkResultsIn` call, the test still passes.  That's because now
the autocomplete does return results... but they don't match the
arbitrary `expected` list that was passed in, so the inner `check`
call fails, and the outer one succeeds because of that exception.

Instead, check more specifically that the MentionAutocompleteView.init
call throws.