autocomplete test: Highlight one end-to-end test, cut the others
It's good to have an end-to-end test here that the actual autocomplete
results reflect the sorting that's tested by all the unit tests in the
rest of this file: a test that makes a MentionAutocompleteView, then
gives it a query, and inspects the results.

The several such test cases that were here, though, don't exercise any
distinct scenarios when it comes to that end-to-end logic; they differ
from each other only in ways that are exercised by the unit tests
above, plus the tests for RecentSenders and RecentDmConversations that
check those data structures' handling of events.  So pick just one of
these test cases and cut the rest.