autocomplete test: Make distinct topic explicitly distinct
This test was relying on its `topic` variable (whose value is
"topic") being different from the topic chosen by default by
eg.streamMessage.  It does happen to be different, but it could
just as well not have been.  Because the difference matters --
the test would fail if they happened to be the same -- the test
case itself should be explicit about that.