message: Expect the absence of newTopic for certain moves.
As noted in:

https://github.com/zulip/zulip-flutter/pull/787#pullrequestreview-2197930076

There is an discrepency between the API docs and the actual server
behavior on the presence of the "new_topic" field on the update message
event. While it was claimed to be always present when a move occurs, it
is actually not the case when the message moves across channels without
chaning its topic.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>