msglist test: Move marker message back and forth.
Previously, messageWithMarker already has its new topic in the message
list when initialized. A subsequent event that moves the message from
elsewhere to said topic does not logically make sense. This fixed that
by moving the message back and forth.

We can't simply move the message from elsewhere because we need to
record the initial rects of the messages.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>