unread_badge_count [nfc]: Use new Figma design variables in neutral variant
These are the same colors in light mode -- with the background color
refactored to match Figma's representation as hex with a percent
opacity -- so this is NFC, since we don't support dark mode yet.

But since these are settled variables in the new Figma, we now have
dark-theme variants for these!