inbox [nfc]: Place provisional dark-theme color variants, pending design
This completes a workable dark-theme implementation for the inbox
screen. It's likely that some colors will be tweaked to follow
designs that haven't been made yet.

As noted in TODOs in each color that's placed here:

- Some of the UI elements are colored with actual design variables
  from the new Figma that's marked ready for dev. But that Figma
  doesn't yet have those same UI elements, so it's unclear if it's
  the right choice of variable.

- Some dark-theme colors were made up by me just now.

NFC because none of the light-theme colors are changed, and we don't
support dark theme yet.