test [nfc]: Remove WidgetTester from testWidgets tests.
This consists of handpicked output from running the dart lint fix
for code `avoid_types_on_closure_parameters`. We probably don't need to
enable this linter rule because in many cases the type annotation
improves readability. For the `(WidgetTester tester)` case, though, it
is common and intuitive enough that the annotation shouldn't be
necessary.

See also:

  https://github.com/zulip/zulip-flutter/pull/762#discussion_r1678487007

Signed-off-by: Zixuan James Li <zixuan@zulip.com>