compose_box [nfc]: Move blockingErrorBanner from _ComposeBoxLayout to caller
This lets _ComposeBoxLayout stay focused on the complexity of how
the text inputs, the editing buttons, and the send button relate
to each other.

In the case where an error banner is present, this also lets us skip
in the build method some logic to compute things that won't end up
getting used.