theme: Follow system setting for light/dark theme!
We now have dark-theme variants in all the styles that differ
between light and dark, in all of the app we've implemented so far.
It's time to let users benefit from that work!

We don't yet let the user choose between dark/light/system in the
app -- that's #97 "Store some client settings" -- but the "system"
behavior should be the default anyway, and this commit provides
that.

The dark-theme colors come from various sources. For each color, I
believe the source is clear either in the code or the Git history.
Those sources are:

- Flutter's library of Material Design widgets, for UI that's been
  using those (like the compose box)
- the "ready for dev" parts of the new Figma, like this:
    https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=2940-48987&m=dev
- the web app (in its state when we wrote the zulip-flutter commit)
- the table of variables in the new Figma, with a note that we're
  not sure if we chose the right variable (because that part of the
  UI doesn't have a "ready for dev" design in the Figma yet)
- my brain and a quick check that it looked OK, as a last resort

We'll want to update the colors as Vlad's work on the new Figma
progresses, and also incorporate his feedback from trying out the
app in dark theme.

Fixes: #95