autocomplete test: Fix up cross-signal tests; add comments to clarify
The way these tests were meant to work wasn't particularly explicit;
and the recent changes adding more signals didn't take that intended
structure into account, so that it became hard to see if the tests
checked what they're intended to check.  Bring them back to that
structure, and add comments explaining it.