actions [nfc]: Factor out `updateMessageFlagsStartingFromAnchor`
Most of the logic in `markNarrowAsRead` is not specific to marking
narrow as read, rather it is actually updating message flags for the
narrow starting from the oldest anchor. As we are going to use this
logic for several other actions starting with the next commit, it is
better to have that generic logic factored out and reused.