actions [nfc]: Simplify name of anchor parameter to update-flags
From the perspective of the callers, this anchor works just like
the "anchor" parameters anywhere else, including the Zulip server
API endpoints that take an anchor and narrow: the function operates
on messages starting from there.  The fact that it does so with
potentially multiple server API calls, each with a different
anchor, is an implementation detail.  So we can just call the
parameter "anchor".