app_bar [nfc]: Explain why the lightbox skips ZulipAppBar
Just listing that there is an exception doesn't help much to provide
guidance for someone writing a new page on whether they should also
skip ZulipAppBar there.  The "why" is much more helpful for that.

Right at the call site is also more helpful than at the definition
of ZulipAppBar, because the call site is the place someone will
definitely be looking if they're contemplating whether this
deviation from the pattern is appropriate, a mistake, or just an
oversight.