api [nfc]: Mark all never-updated User fields final
This was prompted by noticing recently that we weren't updating
[User.isActive]:
  https://github.com/zulip/zulip-flutter/pull/816#discussion_r1689098946

After fixing that, I looked and found there were actually three
other fields on User which we weren't updating -- a latent bug,
since we never looked at those fields, but a bug.  Fixed that in
the preceding commit.

Now the remaining fields that don't get updated are fields that
really are immutable on a Zulip user.  Mark those as final, and
add a comment to help maintain the pattern.