android build: Avoid specifying ndkVersion explicitly
Specifying the NDK version here doesn't seem to do us any good.

This value provided by the `flutter` tool is currently 23.1.7779620.
If we upgrade the Android Gradle plugin (AGP) to 8.5.2, its current
latest, then we start getting build warnings:
  https://github.com/zulip/zulip-flutter/issues/879#issuecomment-2284673367
saying that a bunch of plugins -- including first-party plugins -- are
all using NDK 26.1.10909125 instead, and telling us to switch to that.

Conversely, if we leave this line out, then the same AGP upgrade
causes no warnings and works just fine.

So let the version float.  This does contradict AGP's own recommendation:
  https://developer.android.com/studio/projects/configure-agp-ndk
which is to use this `ndkVersion` property.  But if a bunch of our
plugins are going to use a floating version anyway, it doesn't seem
helpful to try to hang back.

[greg: added commit-message body]