model: Add TypingNotifier
The debugEnable setup is borrowed from
debugEnableRegisterNotificationToken. This acts as a hook to temporarily
disable typing notifications in later widget tests.

The documentation and implementation are based on the web app:

  https://github.com/zulip/zulip/blob/52a9846cdf4abfbe937a94559690d508e95f4065/web/shared/src/typing_status.ts

We unawait setTypingStatus calls.  Because these futures were not
necessary for the rest of the class, we throw them away explicitly
as soon as possible, with unawaited.

Perhaps a nice thing we can do is defining a TypingNotifier mixin and
add that to both PerAccountStore and TypingNotifier, similar to
ChannelStore.  But TypingNotifier is not complicated enough for that to
be worthwhile.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>