msglist: Implement new edited/moved label design.
This replaces the gutter edit state marker design. Since it leads to
message displacement, the test "edit state updates do not affect
layout" is no longer applicable.

We also inline part of what was previously in edit_state_marker.dart,
because the new design simplifies the implementation, and remove the
now unused icons.

The design variables are taken from the Figma design:

  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=3038-56393&t=WMT80mwuFaruNbVf-1

Signed-off-by: Zixuan James Li <zixuan@zulip.com>