deps [nfc]: Update description of Flutter version plans
We don't plan to switch to "beta" or "stable" versions -- better to
live at head, by sticking to the main channel.

(We still want to pin the precise version, but let the issue track
that.  The pubspec here isn't what would control a pin anyway.)

Also point to our instructions for upgrading these versions.