api: Tighten value type on ChannelUpdateEvent for canRemoveSubscribersGroup
This property on streams/channels is documented as type "integer":
  https://zulip.com/api/get-streams#response

(And the same goes for its former name of canRemoveSubscribersGroupId.)

So when the server supports this property at all, and therefore might
send us an event for it, the value it supplies will be non-null.
The corresponding field on ZulipStream is nullable only because of
servers that don't yet support the property.