message: Handle disposal of message store correctly
Part of the work of disposing a MessageListView is to remove it from
a collection of MessageListViews, and we've been doing this in a
loop over the same collection (MessageStoreImpl._messageListViews).
This has been causing concurrent modification errors.

The errors originated from this requirement of dart Iterable:

> Changing a collection while it is being iterated is generally not
  allowed

This fix avoids those errors by having the loop iterate over a clone
of that collection instead.

See also: https://api.flutter.dev/flutter/dart-core/Iterable-class.html

Fixes: #810

Signed-off-by: Zixuan James Li <zixuan@zulip.com>