api: Mark InitialSnapshot.zulipMergeBase as required, relying on server 5+, FL 88+
See "Feature level 88" from Zulip API changelog:

   https://zulip.com/api/changelog

This is one of the few places (along with getServerSettings) where
it's important that we smoothly handle attempting to connect to an
ancient server where this field might indeed be missing.  Happily,
we've already arranged to do that even when the JSON doesn't match
our expectations: see the call site on UpdateMachine of the
constructor ZulipVersionData.fromMalformedServerResponseException.