tools/content: Support systematically surveying unimplemented content features.
We added 2 scripts and a wrapper for them both.

- fetch_messages.dart, the script that fetches messages from a given
  Zulip server, that does not depend on Flutter or other involved
  Zulip Flutter packages, so that it can run without Flutter.  It is
  meant to be run first to produce the corpora needed for surveying
  the unimplemented features.

  The fetched messages are formatted in JSON Lines format, where each
  individual entry is JSON containing the message ID and the rendered
  HTML content.  The script stores output in separate files for
  messages from each server, because message IDs are not unique across
  them.

- unimplemented_features_test.dart, a test that goes over all messages
  collected, parses then with the content parser, and report the
  unimplemented features it discovered.  This is implemented as a test
  mainly because of its dependency on the content parser, which
  depends on the Flutter engine (and `flutter test` conveniently sets
  up a test device).

We mostly avoid prints (https://dart.dev/tools/linter-rules/avoid_print)
in both scripts.  While we don't lose much by disabling this lint rule
for them, because they are supposed to be CLI programs after all, the
rule (potentially) helps with reducing developer inclination to be
verbose.

See comments from the scripts for more details on the implementations.

=====

Some main benefits of having the wrapper script to access dart code
are that we can provide a more intuitive interface consistent with
other tools, for fetching message corpora and/or running the check
for unimplemented features.

Very rarely, you might want to use fetch_messages.dart directly, to use
the `fetch-newer` flag for example to update an existing corpus file.
If we find it helpful, the flag can be added to check-features as well,
but we are skipping that for now.

The script is intended to be run manually, not as a part of the CI,
because it is very slow, and it relies on some out of tree files like
API configs (zuliprc files) and big dumps of chat history.

For the most part, we intend to only keep the detailed explanations in
the underlying scripts close to the implementation, and selectively
repeat some of the helpful information in the wrapper.

This also repeats some easy checks for options, so that we can produce
nicer error messages for some common errors (like missing zuliprc for
`fetch`).

Fixes: #190

Signed-off-by: Zixuan James Li <zixuan@zulip.com>