deps: Upgrade Flutter to 3.25.0-1.0.pre.296
And update Flutter's supporting libraries to match.

Also, as anticipated in #920 (see PR description), handle some
deprecations in the Color API from flutter/engine@eff1b76cf (rolled
in flutter/flutter@8c1a93508). Deprecation warnings are "info"-level
in analyzer output, and CI treats that as fatal, because we don't
want to let those hang around.

In particular, `withOpacity` is replaced with the new `withValues`.

Also, `value` is newly deprecated, but we migrated off that in the
previous commit.