color [nfc]: Add/use extension with `argbInt`, replacing deprecated `value`
The `value` getter is deprecated in flutter/engine@eff1b76cf (rolled
in flutter/flutter@8c1a93508), and I think that's because it could
lose precision for colors in different color spaces that are now
possible to represent. We're about to bump the Flutter minimum
version to take that change.

We're happy with the color space we've been using (sRGB), and this
integer representation has been helpful for us. So, add an extension
for it, allowing consumers to keep using it without each one needing a
`deprecated_member_use`.