msglist: Fix mark-as-read button icon color in dark mode
After sweeping through the message-list screen to implement dark
theme (#843), in late July, the Flutter framework merged
flutter/flutter#143501, in early August. This made it so our
`foregroundColor` no longer controlled the icon color, as promised
in the dartdocs of `foregroundColor` and `iconColor`.

I opened an issue for Flutter about the inconsistency with the doc:
  https://github.com/flutter/flutter/issues/154644
and sent a PR to resolve it, by updating the doc (which the author
of 143501 had said was the right fix):
  https://github.com/flutter/flutter/pull/154646

Fixes: #926