channel [nfc]: Expose debugTopicVisibility directly on store
This follows the pattern we generally use for other members of
ChannelStore and MessageStore, and lets us eliminate the getter
returning a ChannelStoreImpl.

This pattern -- exposing all the data as members directly on the
overall PerAccountStore, delegating to the underlying channel store
and message store as needed -- is convenient because it means that
code consuming the store doesn't have to worry about which sub-area
of the store we've organized a given piece of data into: it's all
just methods on the store.

Noticed this was here while reviewing a PR that included a new
similar deviation from this pattern:
  https://github.com/zulip/zulip-flutter/pull/909#discussion_r1747875935