tools/content: Add wrapper script check-features.
Some main benefits of having a wrapper script to access dart code
are that we can provide a more intuitive interface consistent with
other tools, for fetching message corpuses and/or running the check
for unimplemented features.

Very rarely, you might want to use fetch_messages.dart directly, to use
the `fetch-newer` flag for example to update an existing corpus file.
If we find it helpful, the flag can be added to check-features as well,
but we are skipping that for now.

The script is intended to be run manually, not as a part of the CI,
because it is very slow, and it relies on some out of tree files like
API configs (zuliprc files) and big dumps of chat history.

For the most part, we intend to only keep the detailed explanations in
the underlying scripts close to the implementation, and selectively
repeat some of the helpful information in the wrapper.

This also repeats some easy checks for options, so that we can produce
nicer error messages for some common errors (like missing zuliprc for
`fetch`).

Fixes: #190